+++ services/src/main/resources/META-INF/services/org.keycloak.transaction.JtaTransactionManagerLookup
@@ -1,0 +1,1 @@
+org.keycloak.transaction.JBossJtaTransactionManagerLookup
